Made controller return right responses.

+++ docs_auto_fill/src/main/java/com/logismiko/docs_auto_fill/services/OrganizationService.java
@@ -8,8 +8,11 @@ import com.logismiko.docs_auto_fill.services.factories.OrganizationEntityFactory
 import com.logismiko.docs_auto_fill.services.factories.OrganizationResponseDtoFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+
+import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
 public class OrganizationService {
@@ -18,14 +21,14 @@ public class OrganizationService {
     private OrganizationRepository organizationRepository;
 
     public OrganizationResponseDto addOrganization(OrganizationRequestDto organizationRequestDto) {
-        OrganizationEntity organizationEntity = OrganizationEntityFactory.make(organizationRequestDto);
-        return OrganizationResponseDtoFactory.make(organizationRepository.save(organizationEntity));
+        OrganizationEntity organizationEntity = OrganizationEntityFactory.makeOrganizationEntity(organizationRequestDto);
+        return OrganizationResponseDtoFactory.makeOrganizationResponseDto(organizationRepository.save(organizationEntity));
     }
 
     public OrganizationResponseDto getOrganization(Long id) {
-        return OrganizationResponseDtoFactory.make(
+        return OrganizationResponseDtoFactory.makeOrganizationResponseDto(
                 organizationRepository.findById(id).orElseThrow(
-                        () -> new RuntimeException("Entity: {0} not found".formatted(id))
+                        () -> new ResponseStatusException(NOT_FOUND, String.format("Entity: %s not found", 1L))
                 )
         );
     }
@@ -33,11 +36,18 @@ public class OrganizationService {
     public List<OrganizationResponseDto> getAllOrganizations() {
         return organizationRepository.findAll()
                 .stream()
-                .map(OrganizationResponseDtoFactory::make)
+                .map(OrganizationResponseDtoFactory::makeOrganizationResponseDto)
                 .toList();
     }
 
+    /**
+     * Delete an organization, if it exists.
+     * @param id ID of the organization to be deleted.
+     */
     public void deleteOrganization(Long id) {
-        organizationRepository.deleteById(id);
+        if (organizationRepository.existsById(id))
+            organizationRepository.deleteById(id);
+        else
+            throw new ResponseStatusException(NOT_FOUND, String.format("Entity: %s not found", 1L));
     }
 }
